:rotating_light: Fix 'testpackage'.
