+++ channel/errors_internal_test.go
