+++ channel/persistence/test/peerchans_internal_test.go
