+++ channel/test/app_randomizer_internal_test.go
