+++ client/chanregistry_internal_test.go
