+++ client/proposalopts_internal_test.go
