+++ client/updatemsgs_internal_test.go
