+++ log/log_internal_test.go
