+++ log/none_internal_test.go
