+++ wire/msg_internal_test.go
