+++ wire/net/endpoint_internal_test.go
