+++ wire/net/endpoint_registry_internal_test.go
